+++ include/Font.h
@@ -15,7 +15,7 @@ typedef struct {
     int w;
     int h;
     float scale;
-    float color[4];
+    unsigned char color[4];
 } FontString;
 
 /*!
@@ -35,7 +35,7 @@ public:
 
     virtual void writeString(FontString &s) = 0;
 
-    virtual void drawText(unsigned int x, unsigned int y, float scale, const float color[4], const char *s, ...)
+    virtual void drawText(unsigned int x, unsigned int y, float scale, const unsigned char color[4], const char *s, ...)
         __attribute__((format(printf, 6, 0)));
 
 protected: